+++ grunt/options/watch.js
@@ -6,9 +6,9 @@
 var config = require('../config');
 
 module.exports = {
-  scss: {
-    files: config.sass.files,
-    tasks: 'sass:dev',
+  css: {
+    files: config.css.files,
+    tasks: 'pleeease:dev',
     atBegin: true
   },
 
@@ -21,6 +21,6 @@ module.exports = {
     options: {
       livereload: true
     },
-    files: ['src/css/**/*']
+    files: config.css.files
   }
 };